+++ clients/admin-ui/src/features/data-discovery-and-detection/hooks/useDetectionResultColumns.tsx
@@ -5,8 +5,8 @@ import { RelativeTimestampCell } from "~/features/common/table/v2/cells";
 import DetectionItemAction from "~/features/data-discovery-and-detection/DetectionItemActions";
 import ResultStatusBadgeCell from "~/features/data-discovery-and-detection/tables/ResultStatusBadgeCell";
 import ResultStatusCell from "~/features/data-discovery-and-detection/tables/ResultStatusCell";
-import { DiscoveryMonitorItem } from "~/features/data-discovery-and-detection/types/DiscoveryMonitorItem";
 import { StagedResourceType } from "~/features/data-discovery-and-detection/types/StagedResourceType";
+import { StagedResourceAPIResponse } from "~/types/api";
 
 import findProjectFromUrn from "../utils/findProjectFromUrn";
 
@@ -15,9 +15,9 @@ const useDetectionResultColumns = ({
 }: {
   resourceType?: StagedResourceType;
 }) => {
-  const columnHelper = createColumnHelper<DiscoveryMonitorItem>();
+  const columnHelper = createColumnHelper<StagedResourceAPIResponse>();
 
-  const defaultColumns: ColumnDef<DiscoveryMonitorItem, any>[] = [];
+  const defaultColumns: ColumnDef<StagedResourceAPIResponse, any>[] = [];
 
   if (!resourceType) {
     return { columns: defaultColumns };
@@ -42,10 +42,10 @@ const useDetectionResultColumns = ({
         cell: (props) => <ResultStatusBadgeCell result={props.row.original} />,
         header: (props) => <DefaultHeaderCell value="Status" {...props} />,
       }),
-      columnHelper.display({
-        id: "type",
-        cell: () => <DefaultCell value="Dataset" />,
-        header: "Type",
+      columnHelper.accessor((row) => row.system, {
+        id: "system",
+        cell: (props) => <DefaultCell value={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="System" {...props} />,
       }),
       columnHelper.accessor((row) => row.monitor_config_id, {
         id: "monitor",
@@ -72,6 +72,11 @@ const useDetectionResultColumns = ({
         id: "name",
         cell: (props) => <ResultStatusCell result={props.row.original} />,
         header: (props) => <DefaultHeaderCell value="Table name" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.description, {
+        id: "description",
+        cell: (props) => <DefaultCell value={props.getValue() ?? "--"} />,
+        header: (props) => <DefaultHeaderCell value="Description" {...props} />,
       }),
       columnHelper.display({
         id: "status",
@@ -103,6 +108,11 @@ const useDetectionResultColumns = ({
         id: "name",
         cell: (props) => <ResultStatusCell result={props.row.original} />,
         header: (props) => <DefaultHeaderCell value="Field name" {...props} />,
+      }),
+      columnHelper.accessor((row) => row.description, {
+        id: "description",
+        cell: (props) => <DefaultCell value={props.getValue() ?? "--"} />,
+        header: (props) => <DefaultHeaderCell value="Description" {...props} />,
       }),
       columnHelper.display({
         id: "status",