+++ clients/admin-ui/src/features/data-discovery-and-detection/tables/DetectionResultTable.tsx
@@ -22,10 +22,10 @@ import { useGetMonitorResultsQuery } from "~/features/data-discovery-and-detecti
 import useDetectionResultColumns from "~/features/data-discovery-and-detection/hooks/useDetectionResultColumns";
 import useDiscoveryRoutes from "~/features/data-discovery-and-detection/hooks/useDiscoveryRoutes";
 import IconLegendTooltip from "~/features/data-discovery-and-detection/IndicatorLegend";
-import { DiscoveryMonitorItem } from "~/features/data-discovery-and-detection/types/DiscoveryMonitorItem";
 import { StagedResourceType } from "~/features/data-discovery-and-detection/types/StagedResourceType";
 import { findResourceType } from "~/features/data-discovery-and-detection/utils/findResourceType";
 import getResourceRowName from "~/features/data-discovery-and-detection/utils/getResourceRowName";
+import isNestedField from "~/features/data-discovery-and-detection/utils/isNestedField";
 import { DiffStatus, StagedResource } from "~/types/api";
 
 import { SearchInput } from "../SearchInput";
@@ -115,9 +115,7 @@ const DetectionResultTable = ({ resourceUrn }: MonitorResultTableProps) => {
     search: searchQuery,
   });
 
-  const resourceType = findResourceType(
-    resources?.items[0] as DiscoveryMonitorItem,
-  );
+  const resourceType = findResourceType(resources?.items[0]);
 
   const { columns } = useDetectionResultColumns({ resourceType });
 
@@ -138,14 +136,14 @@ const DetectionResultTable = ({ resourceUrn }: MonitorResultTableProps) => {
 
   const { navigateToDetectionResults } = useDiscoveryRoutes();
 
-  const handleRowClicked =
-    resourceType !== StagedResourceType.FIELD
-      ? (row: StagedResource) =>
-          navigateToDetectionResults({
-            resourceUrn: row.urn,
-            showFullSchema: isShowingFullSchema,
-          })
-      : undefined;
+  const handleRowClicked = (row: StagedResource) =>
+    navigateToDetectionResults({
+      resourceUrn: row.urn,
+      showFullSchema: isShowingFullSchema,
+    });
+
+  const getRowIsClickable = (row: StagedResource) =>
+    resourceType !== StagedResourceType.FIELD || isNestedField(row);
 
   const tableInstance = useReactTable<StagedResource>({
     getCoreRowModel: getCoreRowModel(),
@@ -193,6 +191,7 @@ const DetectionResultTable = ({ resourceUrn }: MonitorResultTableProps) => {
       <FidesTableV2
         tableInstance={tableInstance}
         onRowClick={handleRowClicked}
+        getRowIsClickable={getRowIsClickable}
         emptyTableNotice={<EmptyTableNotice />}
       />
       <PaginationBar