+++ clients/admin-ui/src/pages/data-discovery/activity.tsx
@@ -30,12 +30,7 @@ const DataDiscoveryAndDetectionActivityPage = () => {
         padding: "0 40px 48px",
       }}
     >
-      <PageHeader
-        breadcrumbs={[
-          { title: "Detection & discovery" },
-          { title: "All activity" },
-        ]}
-      />
+      <PageHeader breadcrumbs={[{ title: "All activity" }]} />
       <ActivityTable
         onRowClick={navigateToResourceResults}
         statusFilters={[