+++ src/fides/api/schemas/connection_configuration/connection_secrets_scylla.py
@@ -31,8 +31,7 @@ class ScyllaSchema(ConnectionConfigSecretsSchema):
     )
     keyspace: Optional[str] = Field(
         title="Keyspace",
-        description="The keyspace used.",
-        sensitive=True,
+        description="The keyspace used. If not provided, DSRs for this integration will error. If the integration is used for D & D, then setting a keyspace is not required.",
     )
 
     _required_components: List[str] = [